requirements.txt
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0

psycopg2-binary>=2.9,<3.0


